refactor: extract line.rs
